backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai>=1.0.0
pytest==7.4.3
httpx==0.25.2 